// src/app/shared/interfaces/client.interface.ts
export interface IClient {
    id: number;
    property: number;
    firstName: string;
    middleName: string;
    lastName: string;
    prefix: string;
    suffix: string;
    phoneMobile: number;
    phoneHome: number;
    phoneWork: number;
    emailAddress: string;
    preferredContactType: string;
}